refactor(ReservationHistoryStatus): 계층 변경 및 의존 수정
