+++ src/main/java/roomescape/controller/history/dto/HistoryResponse.java
@@ -1,7 +1,7 @@
 package roomescape.controller.history.dto;
 
 import java.time.LocalDate;
-import roomescape.controller.history.ReservationHistoryStatus;
+import roomescape.domain.history.ReservationHistoryStatus;
 import roomescape.controller.reservationtime.dto.ReservationTimeResponse;
 import roomescape.controller.theme.dto.ThemeResponse;
