+++ src/main/java/roomescape/controller/reservation/ReservationPageModelAssembler.java
@@ -7,12 +7,12 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Component;
 import org.springframework.ui.Model;
-import roomescape.controller.history.ReservationHistoryStatus;
 import roomescape.controller.history.dto.HistoryResponse;
 import roomescape.controller.reservationtime.dto.ReservationTimeResponse;
 import roomescape.controller.reservationtime.dto.ReservationTimeSlotResponse;
 import roomescape.controller.reservationtime.dto.ReservationTimeSlotStatus;
 import roomescape.controller.theme.dto.ThemeResponse;
+import roomescape.domain.history.ReservationHistoryStatus;
 import roomescape.exception.ErrorCode;
 import roomescape.service.history.MyHistoryService;
 import roomescape.service.reservationtime.ReservationTimeService;
