+++ src/main/java/roomescape/domain/history/ReservationHistoryStatus.java
@@ -1,4 +1,4 @@
-package roomescape.controller.history;
+package roomescape.domain.history;
 
 public enum ReservationHistoryStatus {
     RESERVATION,
